requirements.txt
fastapi==0.109.2
slowapi==0.1.8
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlmodel==0.0.18
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==4.1.0
httpx==0.27.0
asgi-lifespan==2.1.0
pydantic~=2.6.4
security~=1.3.1
limiter~=0.5.0